feat(core): get records from store by ID (or all)
