+++ packages/core/src/store.js
@@ -1,13 +1,35 @@
-const data = new WeakMap()
+import mergeOptions from 'merge-options'
+
+const entries = new WeakMap()
 
 class Store {
     constructor() {
-        data.set(this, {})
+        entries.set(this, {})
     }
 
-    add(entry) {
-        const myData = data.get(this)
-        myData[entry.id] = entry
+    addEntry(entry) {
+        entries.get(this)[entry.id] = entry
+    }
+
+    getAll(clone = true) {
+        const entriesMap = entries.get(this)
+        return Object.keys(entriesMap).map(entryId => {
+            if (clone) {
+                return mergeOptions({}, entriesMap[entryId])
+            }
+
+            return entriesMap[entryId]
+        })
+    }
+
+    getById(id, clone = true) {
+        const entry = entries.get(this)[id]
+
+        if (entry && clone) {
+            return mergeOptions({}, entry)
+        }
+
+        return entry || null
     }
 }
