+++ packages/core/test/dummy-plugin.js
@@ -1,0 +1,10 @@
+import {Plugin} from 'core'
+
+class DummyPlugin extends Plugin {
+    constructor(name, loadFunction) {
+        super(name)
+        this.load = loadFunction
+    }
+}
+
+export default DummyPlugin
